refactored timestamp converter to DateTimeConverter
